requirements.txt
# Needed for building the slides
Pygments==2.2.0

# Needed for code samples
Django==2.2.1
django-debug-toolbar==1.11
